removed integration tests and added standalone integration tests to to standalone module
