--- drivers/mongock-driver-mongodb/integration-tests/src/main/java/io/mongock/integrationtests/spring5/springdata3/changelogs/client/ClientUpdater2ChangeLog.java
@@ -1,30 +1,0 @@
-package io.mongock.integrationtests.spring5.springdata3.changelogs.client;
-
-
-import com.github.cloudyrock.mongock.ChangeLog;
-import com.github.cloudyrock.mongock.ChangeSet;
-import com.github.cloudyrock.mongock.driver.mongodb.springdata.v3.decorator.impl.MongockTemplate;
-import io.mongock.integrationtests.spring5.springdata3.Mongock4Spring5SpringData3App;
-import io.mongock.integrationtests.spring5.springdata3.client.ClientExtended;
-
-import java.util.List;
-
-@ChangeLog(order = "3")
-public class ClientUpdater2ChangeLog {
-
-    public final static int INITIAL_CLIENTS = 10;
-
-
-    @ChangeSet(id = "data-updater-with-mongockTemplate-2", order = "001", author = "mongock")
-    public void dataUpdater(MongockTemplate template) {
-
-        List<ClientExtended> clients = template.findAll(ClientExtended.class, Mongock4Spring5SpringData3App.CLIENTS_COLLECTION_NAME);
-
-        clients.stream()
-                .map(client -> client.setName(client.getName() + "_updated_2"))
-                .forEach(client -> template.save(client, Mongock4Spring5SpringData3App.CLIENTS_COLLECTION_NAME));
-
-    }
-
-
-}
