--- drivers/mongock-driver-mongodb/integration-tests/src/main/java/io/mongock/integrationtests/spring5/springdata3/client/ClientRepository.java
@@ -1,10 +1,0 @@
-package io.mongock.integrationtests.spring5.springdata3.client;
-
-import io.mongock.integrationtests.spring5.springdata3.Mongock4Spring5SpringData3App;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository(Mongock4Spring5SpringData3App.CLIENTS_COLLECTION_NAME)
-public interface ClientRepository extends MongoRepository<ClientExtended, String> {
-
-}
