--- drivers/mongock-driver-mongodb/integration-tests/src/main/java/io/mongock/integrationtests/spring5/springdata3/events/MongockFailEventListener.java
@@ -1,13 +1,0 @@
-package io.mongock.integrationtests.spring5.springdata3.events;
-import io.mongock.runner.spring.base.events.SpringMigrationFailureEvent;
-import org.springframework.context.ApplicationListener;
-import org.springframework.stereotype.Component;
-
-@Component
-public class MongockFailEventListener implements ApplicationListener<SpringMigrationFailureEvent> {
-//
-    @Override
-    public void onApplicationEvent(SpringMigrationFailureEvent event) {
-    }
-
-}
