--- drivers/mongock-driver-mongodb/integration-tests/src/main/java/io/mongock/integrationtests/spring5/springdata3/events/MongockStartedEventListener.java
@@ -1,14 +1,0 @@
-package io.mongock.integrationtests.spring5.springdata3.events;
-
-import io.mongock.runner.spring.base.events.SpringMigrationStartedEvent;
-import org.springframework.context.ApplicationListener;
-import org.springframework.stereotype.Component;
-
-@Component
-public class MongockStartedEventListener implements ApplicationListener<SpringMigrationStartedEvent> {
-
-    @Override
-    public void onApplicationEvent(SpringMigrationStartedEvent event) {
-    }
-
-}
