--- drivers/mongock-driver-mongodb/integration-tests/src/main/java/io/mongock/integrationtests/spring5/springdata3/events/MongockSuccessEventListener.java
@@ -1,14 +1,0 @@
-package io.mongock.integrationtests.spring5.springdata3.events;
-
-import io.mongock.runner.spring.base.events.SpringMigrationSuccessEvent;
-import org.springframework.context.ApplicationListener;
-import org.springframework.stereotype.Component;
-
-@Component
-public class MongockSuccessEventListener implements ApplicationListener<SpringMigrationSuccessEvent> {
-
-    @Override
-    public void onApplicationEvent(SpringMigrationSuccessEvent event) {
-    }
-
-}
