--- drivers/mongock-driver-mongodb/integration-tests/src/main/resources/application-dev.yml
@@ -1,5 +1,0 @@
-mongock:
-  migration-scan-package:
-    - io.mongock.integrationtests.spring5.springdata3.changelogs.client
-  trackIgnored: true
-  transactionable: false
