--- drivers/mongock-driver-mongodb/integration-tests/src/test/java/io/mongock/integrationtests/spring5/springdata3/changelogs/empty/EmptyChangeLog.java
@@ -1,12 +1,0 @@
-package io.mongock.integrationtests.spring5.springdata3.changelogs.empty;
-
-import com.github.cloudyrock.mongock.ChangeLog;
-
-/**
- *
- * @since 30.07.14
- */
-@ChangeLog(order = "1")
-public class EmptyChangeLog {
-
-}
