--- drivers/mongock-driver-mongodb/integration-tests/src/test/java/io/mongock/integrationtests/spring5/springdata3/changelogs/transaction/successful/TransactionSuccessfulChangeLog.java
@@ -1,30 +1,0 @@
-package io.mongock.integrationtests.spring5.springdata3.changelogs.transaction.successful;
-
-import com.github.cloudyrock.mongock.ChangeLog;
-import com.github.cloudyrock.mongock.ChangeSet;
-import com.github.cloudyrock.mongock.driver.mongodb.springdata.v3.decorator.impl.MongockTemplate;
-import io.mongock.integrationtests.spring5.springdata3.Mongock4Spring5SpringData3App;
-import io.mongock.integrationtests.spring5.springdata3.client.ClientExtended;
-
-import java.util.Collection;
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.IntStream;
-
-@ChangeLog(order = "2")
-public class TransactionSuccessfulChangeLog {
-
-    @ChangeSet(id = "method-successful", order = "001", author = "mongock")
-    public void methodSuccessful(MongockTemplate template) {
-        List<ClientExtended> clients = IntStream.range(0, 10)
-                .mapToObj(i -> new ClientExtended("name-" + i, "email-" + i, "phone" + i, "country" + i))
-                .collect(Collectors.toList());
-        Collection<ClientExtended> result = template.insert(clients, Mongock4Spring5SpringData3App.CLIENTS_COLLECTION_NAME);
-    }
-
-    @ChangeSet(id = "method-failing", order = "002", author = "mongock")
-    public void methodFailing(MongockTemplate template) {
-    }
-
-
-}
