--- drivers/mongock-driver-mongodb/integration-tests/src/test/java/io/mongock/integrationtests/spring5/springdata3/util/Constants.java
@@ -1,5 +1,0 @@
-package io.mongock.integrationtests.spring5.springdata3.util;
-
-public class Constants {
-    public static final String CHANGELOG_COLLECTION_NAME = "mongockChangeLog";
-}
