+++ mongock-core/mongock-runner/standalone/mongock-standalone/src/test/java/io/mongock/runner/standalone/util/Constants.java
@@ -1,0 +1,6 @@
+package io.mongock.runner.standalone.util;
+
+public class Constants {
+  public static final String CHANGELOG_COLLECTION_NAME = "mongockChangeLog";
+  public static final String DEFAULT_DATABASE_NAME = "test_database";
+}
