+++ mongock-core/mongock-runner/standalone/mongock-standalone/src/test/java/io/mongock/runner/standalone/util/RunnerTestUtil.java
@@ -1,0 +1,25 @@
+package io.mongock.runner.standalone.util;
+
+import com.mongodb.client.MongoClient;
+import io.mongock.driver.mongodb.sync.v4.driver.MongoSync4Driver;
+import io.mongock.runner.standalone.MongockStandalone;
+import io.mongock.runner.standalone.RunnerStandaloneBuilder;
+
+import java.util.Arrays;
+
+public class RunnerTestUtil {
+
+  private final MongoClient mongoClient;
+
+  public RunnerTestUtil(MongoClient mongoClient) {
+    this.mongoClient = mongoClient;
+  }
+
+  public  RunnerStandaloneBuilder getStandaloneBuilderWithMongoDBSync4( String... packagePath) {
+    MongoSync4Driver driver = MongoSync4Driver.withDefaultLock(mongoClient, Constants.DEFAULT_DATABASE_NAME);
+    driver.setChangeLogRepositoryName(Constants.CHANGELOG_COLLECTION_NAME);
+    return  MongockStandalone.builder()
+        .setDriver(driver)
+        .addMigrationScanPackages(Arrays.asList(packagePath));
+  }
+}
