+++ mongock-core/mongock-test-util/src/main/java/io/mongock/util/test/ExpectedException.java
@@ -1,0 +1,9 @@
+package io.mongock.util.test;
+
+public class ExpectedException extends RuntimeException {
+
+  public ExpectedException(String msg) {super(msg);}
+
+
+  public ExpectedException() {super("EXPECTED EXCEPTION");}
+}
